% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_visual.R
\name{as.param}
\alias{as.param}
\title{Constructor of an object "param", which is used for better storage, printing and visualization of the coefficients.}
\usage{
as.param(beta.vec, p, q, m = 1, G.levels = NULL, lambda = NULL)
}
\arguments{
\item{p}{number of levels of T}

\item{q}{number of levels of N}

\item{m}{number of data sets (levels of G)}

\item{G.levels}{levels of group indicator}

\item{lambda}{the tuning parameter associated with this beta, could be a scalar or a vector}

\item{beta.data}{the values of the coefficients. For a single tuning parameter, it is a vector; for multiple
tuning parameters, it is a matrix with each row associated with one tuning parameter}
}
\description{
This function constructs a "param" beta from a vector or a matrix of values of the coefficients. A param object
 can have multiple values of tuning parameter lambda.
}
