% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sim.R
\name{dataset.gen}
\alias{dataset.gen}
\title{Generate one dataset}
\usage{
dataset.gen(beta.mat, p.mat, n, setseed = TRUE, cen.prob = 0)
}
\arguments{
\item{beta.mat}{beta's in a matrix form}

\item{n}{sample size}

\item{setseed}{boolen variable specifying whether fix seed}

\item{cen.prob}{probability of censoring}
}
\value{
data.frame of one dataset
}
\description{
Generate one dataset
}
