% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_visual.R
\name{DoF.param}
\alias{DoF.param}
\title{Function to calculate the degrees of freedom for a "param" beta}
\usage{
DoF.param(beta)
}
\arguments{
\item{beta}{beta of class "param"}
}
\value{
degrees of freedom of each data set, a vector when beta includes one lambda, a matrix when beta
 includes multiple lambda
}
\description{
Function to calculate the degrees of freedom for a "param" beta
}
