% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sol.R
\name{lasso.tree.bic}
\alias{lasso.tree.bic}
\title{Function to select the best beta according to BIC}
\usage{
lasso.tree.bic(G, T, N, y, cen, lambda, inibeta = NULL, trace = TRUE,
  maxiter = 200, eps = 1e-04, w_select = "plainCox", w = NULL)
}
\arguments{
\item{G}{label of dataset}

\item{T}{feature one of the tumor}

\item{N}{feature two of the tumor}

\item{y}{time to failure}

\item{cen}{censor indicator}

\item{lambda}{tuning parameter of the fused group lasso penalty, a vector}

\item{inibeta}{vector of initial values of beta's}

\item{trace}{boolen varable whether to show the process of calculation}

\item{maxiter}{max number of iteration}

\item{eps}{tolerance}

\item{w_select}{the method to select the adaptive weight w, "plainCox" or "preDefine"}

\item{w}{adaptive weight w}
}
\value{
a list including: the best solution of beta according to BIC, a vector of BIC values corresponding
 to the input lambda, and a "param" beta.seq including all the beta's
}
\description{
For a vector of tuning parameters lambda, calculate all the beta and select the best one according to
 BIC criterion
}
\examples{

}
