% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sol.R
\name{logPL}
\alias{logPL}
\title{Function to calculate the negative log partial likelihood}
\usage{
logPL(G, T, N, y, cen, beta)
}
\arguments{
\item{G}{label of dataset}

\item{T, N}{features one of the tumor}

\item{y}{time to failure}

\item{cen}{censor indicator}

\item{beta}{vector of parameter beta's}
}
\value{
the negative log partial likelihood
}
\description{
Function to calculate the negative log partial likelihood
}
