% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MALT.R
\docType{package}
\name{MALT}
\alias{MALT}
\alias{MALT-package}
\title{metaLassoTree: Cancer staging with meta-analysis}
\description{
Categorize cancer patients according to risk, with the Lasso-tree method
 (\url{http://www.ncbi.nlm.nih.gov/pubmed/23907780}), and many similar datasets.
}
\section{Functions}{
 \link{lasso.tree}
}

