% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_visual.R
\name{plot.param}
\alias{plot.param}
\title{Plot function for class "param"}
\usage{
\method{plot}{param}(beta, lambda.index = NULL)
}
\arguments{
\item{beta}{beta is of class "param"}

\item{lambda.index}{when the param beta contains multiple values of the tuning parameter, the user need to
specify the index of the tuning parameter to be plotted}
}
\description{
This function plot the coefficients as an object "param" (see \code{\link{as.param}} for details) on a 2-way
 T/N table, using the darkness of the color to represent the values of coefficients.
}
