% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_visual.R
\name{PlotCurve}
\alias{PlotCurve}
\title{Function to plot the survival curves}
\usage{
PlotCurve(beta, G, T, N, y, cen, ncol = 2, legend.pos = "bottomleft")
}
\arguments{
\item{beta}{the estimated coefficients, in class "param"}

\item{G}{a vector of group labels}

\item{T, N}{cancer characteristic in vectors or factors}

\item{y}{time to events, in vector}

\item{cen}{censor indicator, in vector}

\item{ncol}{number of columns when plotting curves for each trial}

\item{legend.pos}{position of the legend, "bottomleft" or "topright"}
}
\description{
Function to plot the survival curves
}
