% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.malt.R
\name{print.malt}
\alias{print.malt}
\title{print.malt}
\usage{
\method{print}{malt}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{an output of the malt function}

\item{digits}{the number of significative digits}

\item{...}{other parameters}
}
\description{
print.malt
}
