% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_visual.R
\name{similarity}
\alias{similarity}
\title{Function to compare how close two "param" beta's are}
\usage{
similarity(beta0, beta)
}
\arguments{
\item{beta0}{a "param" beta to be compared with, including one lambda value}

\item{beta}{a "param" beta that can have one or multiple lambda values}
}
\value{
a list containing: (1) the similarity between the two beta's, defined by the
 proportion of the correct boundaries; (2) mse, the mean squre error of all the differences
 on the boundary, with repect to beta0; (3) sim.vec, a vector indicating which boundaries
 are the same.
}
\description{
Function to compare how close two "param" beta's are
}
