% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_visual.R
\name{summary.param}
\alias{summary.param}
\title{Summary of the class "param"}
\usage{
\method{summary}{param}(beta, lambda.index = NULL)
}
\arguments{
\item{beta}{an object in class "param"}

\item{lambda.index}{when the param beta contains multiple values of the tuning parameter, the user need to
specify the index of the tuning parameter to be summaried}
}
\description{
This function prints out a "param" object (see \code{\link{as.param}} for details) in a decent way. It only
 works for a param with a single lambda value.
}
