% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_visual.R
\name{tile.plot}
\alias{tile.plot}
\title{Plot the beta matrix as tiles}
\usage{
tile.plot(beta.mat)
}
\arguments{
\item{beta.mat}{the value of beta's in a matrix form}
}
\description{
This function plot the values of beta's in a tile plot: use T and N as axes, and color to represent the values
 of beta's.
}
