% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ess.R
\name{ess}
\alias{ess}
\title{ess}
\usage{
ess(output)
}
\arguments{
\item{output}{an output of the malt function}
}
\value{
a list of effective sample sizes when estimating the first and second moments
}
\description{
computes effective sample sizes of the chain
}
\examples{
d=50
sigma=((d:1)/d)^(1/2)
init=rnorm(d)*sigma
U=function(x){sum(0.5*x^2/sigma^2)}
grad=function(x){x/sigma^2}
n=10^4
g=1.5
h=0.20
L=10
output=malt(init,U,grad,n,g,h,L)
ess(output)
}
