% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_sol.R
\name{lasso.tree}
\alias{lasso.tree}
\title{The core function for lasso-tree with meta-analysis}
\usage{
lasso.tree(G, T, N, y, cen, lambda, inibeta = NULL, trace = TRUE,
  maxiter = 200, eps = 1e-04)
}
\arguments{
\item{G}{label of dataset}

\item{T, N}{features one of the tumor}

\item{y}{time to failure}

\item{cen}{censor indicator}

\item{lambda}{tuning parameter of the fused group lasso penalty, could be a single value or a vector}

\item{inibeta}{vector of initial guess of beta's}

\item{trace}{boolen varable whether to show the process of calculation}

\item{maxiter}{max number of iteration}

\item{eps}{tolerance}
}
\value{
the estimate of lambda in "param" class (see \code{\link{as.param}} for details)
}
\description{
Find the best estimate of parameter beta's, for tuning parameters lambda
}
\examples{

}
