% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/malt.R
\name{malt}
\alias{malt}
\title{malt}
\usage{
malt(potential, grad, n, g, h, T_phys, x_init, warm_up = F)
}
\arguments{
\item{potential}{a function (potential evaluation)}

\item{grad}{a function (gradient evaluation)}

\item{n}{a positive integer (number of MCMC samples)}

\item{g}{a non-negative number (friction)}

\item{h}{a positive number (time step)}

\item{T_phys}{a positive number (integration time)}

\item{x_init}{a vector (initial position)}

\item{warm_up}{a boolean (should the chain be warmed up? default=FALSE). If TRUE then unadjusted trajectories are drawn for n/2 iterations followed by adjusted trajectories for n/2 iterations. The chain starts from the resulting output and the warm up phase is discarded.}
}
\value{
a list with the samples from the chain, the acceptance rate, the effective sample sizes for estimating the first and second moments of each component.
}
\description{
Generates approximate samples from a distribution with density \deqn{\Pi(x)\propto e^{-U(x)}}{Pi(x)=exp(-U(x))/C} Given a potential function \eqn{U}{U} and its gradient evaluation, implements the sampling algorithm: Metropolis Adjusted Langevin Trajectories (Riou-Durand and Vogrinc 2022). Details available at: https://arxiv.org/abs/2202.13230.
}
\examples{
d=50
sigma=((d:1)/d)^(1/2)
potential=function(x){sum(0.5*x^2/sigma^2)}
grad=function(x){x/sigma^2}
n=10^4
g=1.5
h=0.20
T_phys=2
x_init=rnorm(d)*sigma
malt(potential,grad,n,g,h,T_phys,x_init)
}
